+++ src/commands/addhost.ts
@@ -9,7 +9,7 @@ const command: CommandDefinition = {
 	requiredArgs: ["id"],
 	executor: async (msg, args, support) => {
 		const [id] = args;
-		await support.database.authenticateHost(id, msg.author.id, msg.guildID);
+		const tournament = await support.database.authenticateHost(id, msg.author.id, msg.guildID);
 		const newHost = firstMentionOrFail(msg);
 		logger.verbose(
 			JSON.stringify({
@@ -34,7 +34,7 @@ const command: CommandDefinition = {
 				event: "success"
 			})
 		);
-		await reply(msg, `${support.discord.mentionUser(newHost)} added as a host for Tournament ${id}!`);
+		await reply(msg, `${support.discord.mentionUser(newHost)} added as a host for **${tournament.name}**!`);
 	}
 };