+++ test/commands/removechannel.unit.ts
@@ -9,21 +9,21 @@ describe("command:removechannel", function () {
 		msg.channel.createMessage = sinon.spy();
 		await command.executor(msg, ["name"], support);
 		expect(msg.channel.createMessage).to.have.been.calledOnceWithExactly(
-			sinon.match({ content: "This channel removed as a public announcement channel for Tournament name!" })
+			sinon.match({ content: "This channel removed as a public announcement channel for **Tournament 1**!" })
 		);
 	});
 	it("removes public channels", async () => {
 		msg.channel.createMessage = sinon.spy();
 		await command.executor(msg, ["name", "public"], support);
 		expect(msg.channel.createMessage).to.have.been.calledOnceWithExactly(
-			sinon.match({ content: "This channel removed as a public announcement channel for Tournament name!" })
+			sinon.match({ content: "This channel removed as a public announcement channel for **Tournament 1**!" })
 		);
 	});
 	it("removes private channels", async () => {
 		msg.channel.createMessage = sinon.spy();
 		await command.executor(msg, ["name", "private"], support);
 		expect(msg.channel.createMessage).to.have.been.calledOnceWithExactly(
-			sinon.match({ content: "This channel removed as a private announcement channel for Tournament name!" })
+			sinon.match({ content: "This channel removed as a private announcement channel for **Tournament 1**!" })
 		);
 	});
 });